ref(trace-view): Move shared utilities
Ripped from #25264 part 7
