+++ static/app/components/events/interfaces/spans/scrollbarManager.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
+import {clamp, rectOfContent, toPercent} from 'app/components/waterfallTree/utils';
 import getDisplayName from 'app/utils/getDisplayName';
 import {setBodyUserSelect, UserSelectValues} from 'app/utils/userselect';
 
 import {DragManagerChildrenProps} from './dragManager';
-import {clamp, rectOfContent, toPercent} from './utils';
 
 export type ScrollbarManagerChildrenProps = {
   generateContentSpanBarRef: () => (instance: HTMLDivElement | null) => void;
