+++ static/app/views/performance/compare/spanDetail.tsx
@@ -4,9 +4,9 @@ import styled from '@emotion/styled';
 import DateTime from 'app/components/dateTime';
 import {SpanDetailContainer} from 'app/components/events/interfaces/spans/spanDetail';
 import {rawSpanKeys, SpanType} from 'app/components/events/interfaces/spans/types';
-import {getHumanDuration} from 'app/components/events/interfaces/spans/utils';
 import Pill from 'app/components/pill';
 import Pills from 'app/components/pills';
+import {getHumanDuration} from 'app/components/waterfallTree/utils';
 import {t} from 'app/locale';
 import space from 'app/styles/space';
 import getDynamicText from 'app/utils/getDynamicText';
