+++ static/app/views/performance/compare/utils.tsx
@@ -5,8 +5,8 @@ import {
   generateRootSpan,
   isOrphanSpan,
   parseTrace,
-  toPercent,
 } from 'app/components/events/interfaces/spans/utils';
+import {toPercent} from 'app/components/waterfallTree/utils';
 import {EventTransaction} from 'app/types/event';
 
 // Minimum threshold score for descriptions that are similar.
