+++ static/app/views/performance/traceDetails/content.tsx
@@ -23,6 +23,7 @@ import {
   VirtualScrollbar,
   VirtualScrollbarGrip,
 } from 'app/components/waterfallTree/miniHeader';
+import {toPercent} from 'app/components/waterfallTree/utils';
 import {IconInfo} from 'app/icons';
 import {t, tct, tn} from 'app/locale';
 import {Organization} from 'app/types';
@@ -45,7 +46,7 @@ import {
 } from './styles';
 import TransactionGroup from './transactionGroup';
 import {TraceInfo, TreeDepth} from './types';
-import {getTraceInfo, isRootTransaction, toPercent} from './utils';
+import {getTraceInfo, isRootTransaction} from './utils';
 
 type IndexedFusedTransaction = {
   transaction: TraceFullDetailed;
