+++ static/app/views/performance/traceDetails/utils.tsx
@@ -67,8 +67,6 @@ export function isTraceFullDetailed(transaction): transaction is TraceFullDetail
   return Boolean((transaction as TraceFullDetailed).event_id);
 }
 
-export {getHumanDuration, toPercent} from 'app/components/events/interfaces/spans/utils';
-
 export function isRootTransaction(trace: TraceFullDetailed): boolean {
   // Root transactions has no parent_span_id
   return trace.parent_span_id === null;
